requirements.txt
Django==3.0.8
djangorestframework==3.11.0
psycopg2==2.8.5
python-dotenv==0.14.0

